moved options for each dropdown to a different js file and imported into skillform
